requirements.txt
flask
numpy
opencv-python-headless
keras
tensorflow
Pillow
pymongo
flask-bcrypt
flask-login
python-dotenv
kaggle
matplotlib
scikit-learn
pandas
itsdangerous
gunicorn
